middleware pagination refactor
